+++ src/main/kotlin/example/com/model/app_data/AppDataTable.kt
@@ -4,8 +4,8 @@ import org.jetbrains.exposed.sql.Table
 
 object AppDataTable:Table() {
     val appName = varchar("appName",512)
-    val appVersion = integer("appVersion")
-    val appVersionCode = integer("appVersionCode").nullable()
+    val appVersion = double("appVersion")
+    val appVersionCode = double("appVersionCode").nullable()
     val id = integer("id").autoIncrement()
     val appDownloadLink = varchar("appDownloadLink",512)
     val apiKey = varchar("apiKey",512)
